requirements.txt
Flask==3.0.3
numpy==1.24.4
opencv_python==4.12.0.88
photoshop-python-api==0.24.1
pillow==10.4.0
pillow-avif-plugin==1.4.6
pywin32==311
scikit-learn==1.3.2
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
